requirements.txt
discord.py==1.3.0
aiosqlite
httpx
paramiko